Fixed several bugs in the MM code around index and unique index.
That including a bug in unique index where it cannot correctly detect duplicated values, a bug in index model where it failed to sanitize 'key' field to recognize the index build request that should be ignored and so on.

Also enabled unique index generating in correctness to genrate a small portion (5% assuming a uniform distibution of the random int generator) of the indexes as unique indexes.