Remove explicit transactions.
Explicit transactions are disabled for a while. We have bunch of issues
related to connection affinity. They work only with sidecar setup. Even
then, they are generating quiet a few corner cases with schema
management requests.

We have decided to go with MongoDB compatible transactions based on
sessions. Removing this code to make way for them.